pyproject.toml
[project]
name = "dashboard-enem-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
    "pandas-gbq>=0.33.0",
    "streamlit>=1.54.0",
]
